Separate test_pulses into files
